school_gate/requirements.txt
asgiref==3.8.1
Django==5.1.3
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pyzbar==0.1.9
qrcode==8.0
sqlparse==0.5.2
